Move AuthControl outside SsoProcessor
